scripts/pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["reayd-falmouth <dr323090@falmouth.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.1.0"
numpy = "^2.2.3"
scikit-image = "^0.25.2"
svgwrite = "^1.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
